removed flh occurrences, because flh is not possible
