--- s2tbx-s2msi-flhmci-ui/src/main/java/org/esa/s2tbx/processor/flh_mci/docs/package-info.java
@@ -1,4 +1,0 @@
-@HelpSetRegistration(helpSet = "help.hs", position = 4421)
-package org.esa.s2tbx.processor.flh_mci.docs;
-
-import org.netbeans.api.javahelp.HelpSetRegistration;
