--- s2tbx-s2msi-flhmci-ui/src/main/resources/org/esa/s2tbx/processor/flh_mci/docs/flh_mci/FlhMciTool.html
@@ -1,52 +1,0 @@
-<!--
-  ~ Copyright (C) 2010 Brockmann Consult GmbH (info@brockmann-consult.de)
-  ~
-  ~ This program is free software; you can redistribute it and/or modify it
-  ~ under the terms of the GNU General Public License as published by the Free
-  ~ Software Foundation; either version 3 of the License, or (at your option)
-  ~ any later version.
-  ~ This program is distributed in the hope that it will be useful, but WITHOUT
-  ~ ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-  ~ FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
-  ~ more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License along
-  ~ with this program; if not, see http://www.gnu.org/licenses/
-  -->
-
-<html>
-<head>
-    <title>SNAP Data Processors - S2 FLH/MCI Processor</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1">
-    <link rel="stylesheet" href="../style.css">
-</head>
-
-<body>
-<table class="header">
-    <tr class="header">
-        <td class="header">&nbsp;
-            S2 FLH/MCI Processor</td>
-        <td class="header" align="right"><a href="nbdocs://org.esa.snap.snap.help/org/esa/snap/snap/help/docs/general/overview/SnapOverview.html"><img src="../images/snap_header.jpg" border=0></a>
-        </td>
-    </tr>
-</table>
-
-<p>
-    The S2 FLH/MCI Processor calculates the <u><b>F</b></u>luorescence
-    <u><b>L</b></u>ine <u><b>H</b></u>eight and the <u><b>M</b></u>aximum
-    <u><b>C</b></u>hlorophyll <b><u>I</u></b>ndex by exploiting the height
-    of a measurement over a specific baseline.
-</p>
-
-<p>
-    The Processor can also be used to define new line height products because both
-    the baseline bands and the signal band are freely configurable.
-</p>
-
-<p><a href="FlhMciAlgorithmSpecification.html">S2 FLH/MCI Algorithm Specification </a></p>
-
-<p><a href="FlhMciProcessorDescription.html">S2 FLH/MCI Processor Documentation</a></p>
-<br>
-<hr>
-</body>
-</html>
