--- s2tbx-s2msi-flhmci-ui/src/main/resources/org/esa/s2tbx/processor/flh_mci/docs/help.hs
@@ -1,24 +1,0 @@
-<?xml version='1.0' encoding='ISO-8859-1' ?>
-<!DOCTYPE helpset
-        PUBLIC "-//Sun Microsystems Inc.//DTD JavaHelp HelpSet Version 2.0//EN"
-        "http://java.sun.com/products/javahelp/helpset_2_0.dtd">
-
-<helpset version="2.0">
-    <title>S2 FlhMci Processor Help</title>
-    <maps>
-        <homeID>top</homeID>
-        <mapref location="map.jhm"/>
-    </maps>
-    <view mergetype="javax.help.UniteAppendMerge">
-        <name>TOC</name>
-        <label>Contents</label>
-        <type>javax.help.TOCView</type>
-        <data>toc.xml</data>
-    </view>
-    <view>
-        <name>Search</name>
-        <label>Search</label>
-        <type>javax.help.SearchView</type>
-        <data engine="com.sun.java.help.search.DefaultSearchEngine">JavaHelpSearch</data>
-    </view>
-</helpset>
