--- s2tbx-s2msi-flhmci-ui/src/main/resources/org/esa/s2tbx/processor/flh_mci/docs/map.jhm
@@ -1,12 +1,0 @@
-<?xml version='1.0' encoding='ISO-8859-1' ?>
-<!DOCTYPE map
-        PUBLIC "-//Sun Microsystems Inc.//DTD JavaHelp Map Version 2.0//EN"
-        "http://java.sun.com/products/javahelp/map_2_0.dtd">
-
-<map version="2.0">
-
-    <mapID target="s2flhMciScientificTool" url="flh_mci/FlhMciTool.html"/>
-    <mapID target="s2flhMciAlgorithmSpec" url="flh_mci/FlhMciAlgorithmSpecification.html"/>
-    <mapID target="s2flhMciProcessorDesc" url="flh_mci/FlhMciProcessorDescription.html"/>
-
-</map>
