--- s2tbx-s2msi-flhmci-ui/src/main/resources/org/esa/s2tbx/processor/flh_mci/docs/toc.xml
@@ -1,20 +1,0 @@
-<?xml version='1.0' encoding='ISO-8859-1'  ?>
-<!DOCTYPE toc
-        PUBLIC
-        "-//Sun Microsystems Inc.//DTD
-     JavaHelp TOC Version 2.0//EN"
-        "http://java.sun.com/products/javahelp/toc_2_0.dtd">
-
-<toc version="2.0" categoryclosedimage="folder" categoryopenimage="folder" topicimage="doc">
-
-    <tocitem text="Sentinel Toolbox Application">
-        <tocitem text="Optical">
-            <tocitem text="Thematic Water Processing">
-                <tocitem text="S2 FLH/MCI Processor" target="s2flhMciScientificTool">
-                    <tocitem text="Algorithm Specification" target="s2flhMciAlgorithmSpec"/>
-                    <tocitem text="Processor Description" target="s2flhMciProcessorDesc"/>
-                </tocitem>
-            </tocitem>
-        </tocitem>
-    </tocitem>
-</toc>
