--- s2tbx-s2msi-flhmci/src/main/java/org/esa/s2tbx/processor/flh_mci/package-info.java
@@ -1,5 +1,0 @@
-/**
- * This package provides an implementation for a Fluorescence Line Height (FLH)
- * and Maximum Chlorophyl Index (MCI) Processor for ENVISAT MERIS products.
- */
-package org.esa.s3tbx.processor.flh_mci;
