--- s2tbx-s2msi-flhmci/src/main/nbm/manifest.mf
@@ -1,15 +1,0 @@
-Manifest-Version: 1.0
-AutoUpdate-Show-In-Client: false
-AutoUpdate-Essential-Module: true
-OpenIDE-Module-Java-Dependencies: Java > 1.8
-OpenIDE-Module-Display-Category: Sentinel-3 Toolbox
-OpenIDE-Module-Specification-Version: ${s3tbx.nbmSpecVersion}
-OpenIDE-Module-Layer: org/esa/s3tbx/processor/flh_mci/layer.xml
-OpenIDE-Module-Long-Description: <p>The MERIS FLH/MCI Processor calcul
- ates the Fluorescence Line Height and the Maximum Chlorophyll Index b
- y exploiting the height of a measurement over a specific baseline.</p
- ><p><b>Vendor:</b> Brockmann Consult GmbH</p><p><b>Contact address:</
- b> Max-Planck-Str. 2, 21502 Geesthacht (Germany)</p><p><b>Copyright:<
- /b> (C) 2014 by Brockmann Consult GmbH</p><p><b>Vendor:</b> Brockmann
-  Consult GmbH</p><p><b>License:</b> http://www.gnu.org/licenses/gpl.h
- tml</p>
