--- s2tbx-s2msi-flhmci/src/main/resources/org/esa/s2tbx/processor/flh_mci/layer.xml
@@ -1,10 +1,0 @@
-<filesystem>
-
-    <folder name="OperatorUIs">
-        <file name="org.esa.s2tbx.processor.flh_mci">
-            <attr name="operatorName" stringvalue="FlhMci.s2"/>
-            <attr name="disableFromGraphBuilder" stringvalue="true"/>
-        </file>
-    </folder>
-
-</filesystem>
