+++ s2tbx-s2msi-mci-ui/src/main/java/org/esa/s2tbx/processor/mci/docs/package-info.java
@@ -1,0 +1,4 @@
+@HelpSetRegistration(helpSet = "help.hs", position = 4421)
+package org.esa.s2tbx.processor.mci.docs;
+
+import org.netbeans.api.javahelp.HelpSetRegistration;
