+++ s2tbx-s2msi-mci-ui/src/main/nbm/manifest.mf
