+++ s2tbx-s2msi-mci-ui/src/main/resources/org/esa/s2tbx/processor/mci/docs/map.jhm
@@ -1,0 +1,10 @@
+<?xml version='1.0' encoding='ISO-8859-1' ?>
+<!DOCTYPE map PUBLIC "-//Sun Microsystems Inc.//DTD JavaHelp Map Version 2.0//EN" "http://java.sun.com/products/javahelp/map_2_0.dtd">
+
+<map version="2.0">
+
+    <mapID target="s2MciScientificTool" url="mci/MciTool.html"/>
+    <mapID target="s2MciAlgorithmSpec" url="mci/MciAlgorithmSpecification.html"/>
+    <mapID target="s2MciProcessorDesc" url="mci/MciProcessorDescription.html"/>
+
+</map>
