+++ s2tbx-s2msi-mci-ui/src/main/resources/org/esa/s2tbx/processor/mci/docs/mci/MciProcessorDescription.html
@@ -1,0 +1,183 @@
+<!--
+  ~ Copyright (C) 2010 Brockmann Consult GmbH (info@brockmann-consult.de)
+  ~
+  ~ This program is free software; you can redistribute it and/or modify it
+  ~ under the terms of the GNU General Public License as published by the Free
+  ~ Software Foundation; either version 3 of the License, or (at your option)
+  ~ any later version.
+  ~ This program is distributed in the hope that it will be useful, but WITHOUT
+  ~ ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+  ~ FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+  ~ more details.
+  ~
+  ~ You should have received a copy of the GNU General Public License along
+  ~ with this program; if not, see http://www.gnu.org/licenses/
+  -->
+
+<html>
+<head>
+    <title>SNAP Data Processors - S2 FLH/MCI</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=iso-8859-1">
+    <link rel="stylesheet" href="../style.css">
+</head>
+
+<body>
+<table class="header">
+    <tr class="header">
+        <td class="header">&nbsp;
+            Processor Description - S2 FLH/MCI
+        </td>
+        <td class="header" align="right"><a href="nbdocs://org.esa.snap.snap.help/org/esa/snap/snap/help/docs/general/overview/SnapOverview.html"><img src="../images/snap_header.jpg"
+                                                                                     border=0></a>
+        </td>
+    </tr>
+</table>
+
+<h3>User Interface</h3>
+
+<p>
+    The S2 FLH/MCI processor can be invoked in the Sentinel Toolbox from tool menu by selecting
+    <b>Processing->Thematic Water Processing->S2 FLH/MCI Processor...</b>.
+    On the command line the FLH/MCI processor is available by means of the Graph Processing Tool <code>gpt</code>
+    which is located in the Sentinel Toolbox <code>bin</code> directory.
+    Typing <code>gpt FlhMci.s2 -h</code> displays further information.
+</p>
+
+<p>
+    Selecting the <b>S2 FLH/MCI Processor</b> command from the Sentinel Toolbox tool menu pops up the following dialog:
+</p>
+
+<p align="center"><img src="images/FlhPaths.png"></p>
+
+<h5>Source Product Group</h5>
+
+<p class="i1">
+    <b>Source product:</b>
+    Here the user specifies the source product.
+    The combo box presents a list of all products opened in the Sentinel Toolbox.
+    The user may select one of these or, by clicking on the button next to the combo box, choose a product
+    from the file system.
+</p>
+
+<h5>Target Product Group</h5>
+
+<p class="i1">
+    <b>Name:</b>
+    Used to specify the name of the target product.
+</p>
+
+<p class="i1">
+    <b>Save as:</b>
+    Used to specify whether the target product should be saved to the file system. The
+    combo box presents a list of available file formats.
+    The text field or the button next to it allow to specify a target directory.
+</p>
+
+<p class="i1">
+    <b>Open in SNAP:</b>
+    Used to specify whether the target product should be opened in the Sentinel Toolbox.
+    When the target product is not saved, it is opened in the Sentinel Toolbox automatically.
+</p>
+
+<p align="center"><img src="images/FlhParameter.png"></p>
+
+<h5>Processing Parameters Panel</h5>
+
+<p class="i1">
+    <b>Preset:</b> Select one of the available presets from the predefined list.
+    Available presets are:
+</p>
+<dl class="i1">
+    <dt>S2 MSI</dt>
+    <dd>S2 MSI maximum chlorophyll index (lower=radiance_8, upper=radiance_10, signla=radiance_9)</dd>
+    <dt>S2 FLH</dt>
+    <dd>S2 MSI product fluorescence line height(lower=reflec_7, upper=reflec_9, signla=reflec_8)</dd>
+</dl>
+<p class="i1">
+    A preset selects the bands used to calculate the baseline, the signal band,
+    the line height band and the slope band.
+</p>
+
+<p class="i1">
+    <b>Lower baseline band name:</b> The name for the lower wavelength band defining
+    the baseline. Enter a band name or select one of the list of available bands
+    (only available when an input product is already selected).
+</p>
+
+<p class="i1">
+    <b>Upper baseline band name:</b>
+    The name of the upper wavelength band defining
+    the baseline. Enter a band name or select one of the list of available bands
+    (only available when an input product is already selected).
+</p>
+
+<p class="i1">
+    <b>Signal band name:</b> The name of the signal band, i.e. the band for
+    which the baseline height is calculated. Enter a band name or select one of
+    the list of available bands (only available when an input product is already
+    selected).
+</p>
+
+<p class="i1">
+    <b>Line height band name:</b> The name of the line height band in the
+    target product.
+</p>
+
+<p class="i1">
+    <b>Generate slope parameter:</b>
+    Activates or deactivates calculating the slope parameter.
+</p>
+
+<p class="i1">
+    <b>Slope band name:</b> The name of the slope band in the target product
+    (disabled when calculating the slope is deactivated)
+</p>
+
+<p class="i1">
+    <b>Mask expression:</b>
+    A ROI-mask expression to be evaluated during processing. Used to define pixels of interest.
+</p>
+
+<p class="i1">
+    <b>Cloud correction factor:</b>
+    The cloud correction factor used during calculation. See <a href="MciAlgorithmSpecification.html">algorithm
+    documentation</a> for details.
+</p>
+
+<p class="i1">
+    <b>Invalid FLH/MCI value:</b>
+    The value written to a target pixel when it is excluded from the ROI or the FLH/MCI calculation yields an invalid
+    result.
+</p>
+
+<h5>Button Group</h5>
+
+<p class="i1">
+    <b>Run</b>
+    Creates the target product. The FLH/MCI processing is actually deferred until its band data are accessed, either by
+    writing the product to a file or by viewing its band data. When the <b>Save as</b> option is checked, the FLH/MCI
+    processing is triggered automatically.
+</p>
+
+<p class="i1">
+    <b>Close</b>
+    Closes the dialog.
+</p>
+
+<h5>Menu Group</h5>
+
+<p class="i1">
+    <b>File</b>
+    The entries found in this menu allow to save the current processing parameters to disk and to open a saved parameter
+    set from disk.
+</p>
+
+<p class="i1">
+    <b>Help</b>
+    Displays this help page.
+</p>
+
+<br>
+<hr>
+</body>
+</html>
