+++ s2tbx-s2msi-mci-ui/src/main/resources/org/esa/s2tbx/processor/mci/docs/toc.xml
@@ -1,0 +1,16 @@
+<?xml version='1.0' encoding='ISO-8859-1'  ?>
+<!DOCTYPE toc PUBLIC "-//Sun Microsystems Inc.//DTD JavaHelp TOC Version 2.0//EN" "http://java.sun.com/products/javahelp/toc_2_0.dtd">
+
+<toc version="2.0" categoryclosedimage="folder" categoryopenimage="folder" topicimage="doc">
+
+    <tocitem text="Sentinel Toolbox Application">
+        <tocitem text="Optical">
+            <tocitem text="Thematic Water Processing">
+                <tocitem text="S2 MCI Processor" target="s2MciScientificTool">
+                    <tocitem text="Algorithm Specification" target="s2MciAlgorithmSpec"/>
+                    <tocitem text="Processor Description" target="s2MciProcessorDesc"/>
+                </tocitem>
+            </tocitem>
+        </tocitem>
+    </tocitem>
+</toc>
