+++ s2tbx-s2msi-mci/src/main/java/org/esa/s2tbx/processor/mci/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides an implementation for a Maximum Chlorophyl Index (MCI) Processor for Sentinel-2 MSI data products.
+ */
+package org.esa.s2tbx.processor.mci;
