+++ s2tbx-s2msi-mci/src/main/nbm/manifest.mf
@@ -1,0 +1,7 @@
+Manifest-Version: 1.0
+AutoUpdate-Show-In-Client: false
+AutoUpdate-Essential-Module: true
+OpenIDE-Module-Java-Dependencies: Java > 1.8
+OpenIDE-Module-Display-Category: Sentinel-3 Toolbox
+OpenIDE-Module-Specification-Version: ${s2tbx.nbmSpecVersion}
+OpenIDE-Module-Layer: org/esa/s2tbx/processor/flh_mci/layer.xml
