+++ s2tbx-s2msi-mci/src/main/resources/META-INF/services/org.esa.snap.core.gpf.OperatorSpi
